// src/modules/listings/dto/nutritions.dto.ts
export class GetNutritionsResponse {
    calories: {value: number, unit: string};
    total_fat: {value: number, unit: string};
    saturated_fat: {value: number, unit: string};
    cholesterol: {value: number, unit: string};
    sodium: {value: number, unit: string};
    total_carbohydrate: {value: number, unit: string};
    dietary_fiber: {value: number, unit: string};
    sugars: {value: number, unit: string};
    protein: {value: number, unit: string};
    potassium: {value: number, unit: string};
}